fix dot cloud reader bug unable to read file

+++ Source/DotCloudReader.cpp
@@ -18,14 +18,14 @@ vector<Vector3D*> DotCloudReader::GetDotCloud()
     ifstream file(filename);
 
     double x = 0, y = 0, z = 0;
-    uint8_t red = 0, green = 0, blue = 0;
+    int red = 0, green = 0, blue = 0;
     char hex;
 
     //each row start with a "#"
     while (file >> hex)
     {
         file >> x >> y >> z >> red >> green >> blue;
-        Vector3D* dot = new Vector3D(x, y, z, red, green, blue);
+        Vector3D* dot = new Vector3D(x, y, z, (uint8_t)red, (uint8_t)green, (uint8_t)blue);
         dots.push_back(dot);
     }
